+++ apps/wallets/smart-wallet/react/src/app/utils/createAAWallet/createPasskeyWallet.ts
@@ -1,6 +1,6 @@
-import { Blockchain, SmartWalletSDK } from "@crossmint/client-sdk-aa-passkeys-beta";
+import { Chain, SmartWalletSDK } from "@crossmint/client-sdk-smart-wallet";
 
-import { checkAuthState, parseToken, signInWithGoogle } from "../../auth/FirebaseAuthManager";
+import { checkAuthState, signInWithGoogle } from "../../auth/FirebaseAuthManager";
 
 export async function createPasskeyWallet(isProd: boolean) {
     let jwt = await checkAuthState();
@@ -13,8 +13,6 @@ export async function createPasskeyWallet(isProd: boolean) {
         throw new Error("No JWT token found");
     }
 
-    const { sub } = parseToken(jwt);
-
     const xm = isProd
         ? SmartWalletSDK.init({
               clientApiKey: process.env.REACT_APP_CROSSMINT_API_KEY_PROD || "",
@@ -23,7 +21,7 @@ export async function createPasskeyWallet(isProd: boolean) {
               clientApiKey: process.env.REACT_APP_CROSSMINT_API_KEY_STG || "",
           });
 
-    const chain = isProd ? Blockchain.POLYGON : Blockchain.POLYGON_AMOY;
+    const chain = isProd ? Chain.POLYGON : Chain.POLYGON_AMOY;
 
     const test = await xm.getOrCreateWallet({ jwt }, chain);