+++ packages/client/wallets/smart-wallet/src/blockchain/wallets/paymaster.ts
@@ -1,39 +1,36 @@
-import { PAYMASTER_RPC } from "@/utils/constants";
-import { createZeroDevPaymasterClient } from "@zerodev/sdk";
+import { CrossmintWalletService } from "@/api/CrossmintWalletService";
+import { UserParams } from "@/types/Config";
 import { Middleware } from "permissionless/actions/smartAccount";
 import { EntryPoint } from "permissionless/types/entrypoint";
-import { http } from "viem";
 
 import { usesGelatoBundler } from "../../utils/blockchain";
-import { SmartWalletChain, viemNetworks, zerodevProjects } from "../chains";
+import { SmartWalletChain } from "../chains";
 
 export function usePaymaster(chain: SmartWalletChain) {
     return !usesGelatoBundler(chain);
 }
 
-const getPaymasterRPC = (chain: SmartWalletChain) => {
-    return PAYMASTER_RPC + zerodevProjects[chain];
-};
-
 export function paymasterMiddleware({
     entryPoint,
     chain,
+    walletService,
+    user,
 }: {
     entryPoint: EntryPoint;
     chain: SmartWalletChain;
+    walletService: CrossmintWalletService;
+    user: UserParams;
 }): Middleware<EntryPoint> {
     return {
         middleware: {
             sponsorUserOperation: async ({ userOperation }) => {
-                const paymasterClient = createZeroDevPaymasterClient({
-                    chain: viemNetworks[chain],
-                    transport: http(getPaymasterRPC(chain)),
-                    entryPoint,
-                });
-                return paymasterClient.sponsorUserOperation({
+                const { sponsorUserOpParams } = await walletService.sponsorUserOperation(
+                    user,
                     userOperation,
                     entryPoint,
-                });
+                    chain
+                );
+                return sponsorUserOpParams;
             },
         },
     };