+++ packages/client/wallets/smart-wallet/src/utils/blockchain.ts
@@ -1,4 +1,4 @@
-import { SmartWalletChain } from "..";
+import { SmartWalletChain } from "../blockchain/chains";
 
 export function usesGelatoBundler(chain: SmartWalletChain) {
     return false;