+++ packages/client/auth-core/CHANGELOG.md
@@ -1,5 +1,18 @@
 # @crossmint/client-sdk-auth-core
 
+## 1.0.2
+
+### Patch Changes
+
+-   Updated dependencies [11a9bb8]
+    -   @crossmint/client-sdk-window@0.0.9
+
+## 1.0.1
+
+### Patch Changes
+
+-   82e429b: Make the package public
+
 ## 1.0.0
 
 ### Major Changes